数据树功能开发

+++ 后台管理系统/vue-managementsystem-web/src/moduleA/pc/components/home/tables/UserMessages.vue
@@ -116,13 +116,16 @@
 
       // 批量删除
       deleteList: function () {
+        this.ids = new Array();
+        for (let item of this.multipleSelection) {
+          this.ids.push(item.leaveMsgId);
+        }
+        if (Tools.isNull(this.ids)) {
+          return;
+        }
         this.messageBox.confirm(this.$t('rs.staticText.30000000010'), this.$t('rs.staticText.30000000008'), () => {
         }, () => {
           // 确定
-          this.ids = new Array();
-          for (let item of this.multipleSelection) {
-            this.ids.push(item.leaveMsgId);
-          }
           this.delete();
         }, () => {
           // 取消
